Init v1
Initial upload.

This repo is a personal project focused on creating a card game similar to Magic: The Gathering. The theme is a sci-fi card game, in which the players incubate and fight each other with viruses that they play onto the battlefield. They protect themselves with the positive effects created by implementing Firewall Modules into their defensive firewalls.

TODO:
- Add/handle card play locations on table
- Add player System Totals
- Add GUI
- Create Main Menu
- Create In-Game Menu
- Create In-Game Chat/Emojis/Stickers?
- Create Buddy System?
- Implement Steam/Epic Networking?

COMPLETED:
- Take a deck list and check it against a card database. If all cards are valid, generate an array of cards for that deck and shuffle them.
- Drawing cards
- Displaying cards in hand, complete with sorting and interpolated movement
- In terms of replication, each player has their own copy of all cards on the server, and the server is the authority in actions such as drawing cards.
- Display basic information on player cards